added new class to reconfigure stufff
