+++ src/Features/Core/Portable/IntroduceVariable/AbstractIntroduceParameterService.cs
@@ -5,16 +5,13 @@
 using System;
 using System.Collections.Generic;
 using System.Collections.Immutable;
-using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.CodeAnalysis.CodeActions;
-using Microsoft.CodeAnalysis.CodeGeneration;
 using Microsoft.CodeAnalysis.CodeRefactorings;
 using Microsoft.CodeAnalysis.Editing;
 using Microsoft.CodeAnalysis.FindSymbols;
-using Microsoft.CodeAnalysis.Formatting;
 using Microsoft.CodeAnalysis.LanguageServices;
 using Microsoft.CodeAnalysis.Shared.Collections;
 using Microsoft.CodeAnalysis.Shared.Extensions;
@@ -34,8 +31,8 @@ namespace Microsoft.CodeAnalysis.IntroduceVariable
         where TIdentifierNameSyntax : TExpressionSyntax
     {
         protected abstract SyntaxNode GenerateExpressionFromOptionalParameter(IParameterSymbol parameterSymbol);
-        protected abstract SyntaxNode? GetLocalDeclarationFromDeclarator(SyntaxNode variableDecl);
         protected abstract SyntaxNode UpdateArgumentListSyntax(SyntaxNode argumentList, SeparatedSyntaxList<SyntaxNode> arguments);
+        protected abstract SyntaxNode? GetLocalDeclarationFromDeclarator(SyntaxNode variableDecl);
         protected abstract bool IsDestructor(IMethodSymbol methodSymbol);
 
         public sealed override async Task ComputeRefactoringsAsync(CodeRefactoringContext context)
@@ -219,92 +216,36 @@ namespace Microsoft.CodeAnalysis.IntroduceVariable
             IMethodSymbol methodSymbol, SyntaxNode containingMethod, bool allOccurrences, IntroduceParameterCodeActionKind selectedCodeAction,
             CancellationToken cancellationToken)
         {
-            var parameterName = await GetNewParameterNameAsync(originalDocument, expression, cancellationToken).ConfigureAwait(false);
-
             var methodCallSites = await FindCallSitesAsync(originalDocument, methodSymbol, cancellationToken).ConfigureAwait(false);
 
             var modifiedSolution = originalDocument.Project.Solution;
-            var mappingDictionary = await MapExpressionToParametersAsync(originalDocument, expression, cancellationToken).ConfigureAwait(false);
             var syntaxFacts = originalDocument.GetRequiredLanguageService<ISyntaxFactsService>();
+            var introduceParameterState = new IntroduceParameterDocumentRewriter(this, originalDocument, expression, methodSymbol, containingMethod, selectedCodeAction, allOccurrences);
 
             foreach (var (project, projectCallSites) in methodCallSites.GroupBy(kvp => kvp.Key.Project))
             {
                 var compilation = await project.GetRequiredCompilationAsync(cancellationToken).ConfigureAwait(false);
                 foreach (var (document, invocations) in projectCallSites)
                 {
-                    var insertionIndex = GetInsertionIndex(compilation, methodSymbol, syntaxFacts, containingMethod);
-                    if (selectedCodeAction is IntroduceParameterCodeActionKind.Trampoline or IntroduceParameterCodeActionKind.Overload)
+                    var newRoot = await introduceParameterState.ModifyDocumentAsync(compilation, document, invocations, cancellationToken).ConfigureAwait(false);
+                    modifiedSolution = modifiedSolution.WithDocumentSyntaxRoot(originalDocument.Id, newRoot);
+                    /*if (selectedCodeAction is IntroduceParameterCodeActionKind.Trampoline or IntroduceParameterCodeActionKind.Overload)
                     {
-                        var newRoot = await ModifyDocumentInvocationsTrampolineOverloadAndIntroduceParameterAsync(compilation,
-                            originalDocument, document, invocations, mappingDictionary, methodSymbol, containingMethod,
-                            insertionIndex, allOccurrences, parameterName, expression, selectedCodeAction, cancellationToken).ConfigureAwait(false);
+
+                        var newRoot = await ModifyDocumentInvocationsTrampolineOverloadAndIntroduceParameterAsync(introduceParameterState,
+                            compilation, document, invocations, insertionIndex).ConfigureAwait(false);
                         modifiedSolution = modifiedSolution.WithDocumentSyntaxRoot(originalDocument.Id, newRoot);
                     }
                     else
                     {
-                        var newRoot = await ModifyDocumentInvocationsAndIntroduceParameterAsync(compilation,
-                            originalDocument, document, mappingDictionary, containingMethod,
-                            expression, allOccurrences, parameterName, insertionIndex, invocations,
-                            cancellationToken).ConfigureAwait(false);
+                        var newRoot = await ModifyDocumentInvocationsAndIntroduceParameterAsync(introduceParameterState,
+                            compilation, document, insertionIndex, invocations).ConfigureAwait(false);
                         modifiedSolution = modifiedSolution.WithDocumentSyntaxRoot(originalDocument.Id, newRoot);
-                    }
+                    }*/
                 }
             }
 
             return modifiedSolution;
-        }
-
-        /// <summary>
-        /// Gets the parameter name, if the expression's grandparent is a variable declarator then it just gets the
-        /// local declarations name. Otherwise, it generates a name based on the context of the expression.
-        /// </summary>
-        private async Task<string> GetNewParameterNameAsync(Document document, TExpressionSyntax expression, CancellationToken cancellationToken)
-        {
-            if (ShouldRemoveVariableDeclaratorContainingExpression(document, expression, out var varDeclName, out _))
-            {
-                return varDeclName;
-            }
-
-            var semanticModel = await document.GetRequiredSemanticModelAsync(cancellationToken).ConfigureAwait(false);
-            var semanticFacts = document.GetRequiredLanguageService<ISemanticFactsService>();
-            return semanticFacts.GenerateNameForExpression(semanticModel, expression, capitalize: false, cancellationToken);
-        }
-
-        /// <summary>
-        /// Determines if the expression's grandparent is a variable declarator and if so,
-        /// returns the name
-        /// </summary>
-        protected bool ShouldRemoveVariableDeclaratorContainingExpression(
-            Document document, TExpressionSyntax expression, [NotNullWhen(true)] out string? varDeclName, [NotNullWhen(true)] out SyntaxNode? localDeclaration)
-        {
-            var syntaxFacts = document.GetRequiredLanguageService<ISyntaxFactsService>();
-            var declarator = expression?.Parent?.Parent;
-
-            localDeclaration = null;
-
-            if (!syntaxFacts.IsVariableDeclarator(declarator))
-            {
-                varDeclName = null;
-                return false;
-            }
-
-            localDeclaration = GetLocalDeclarationFromDeclarator(declarator);
-            if (localDeclaration is null)
-            {
-                varDeclName = null;
-                return false;
-            }
-
-            // TODO: handle in the future
-            if (syntaxFacts.GetVariablesOfLocalDeclarationStatement(localDeclaration).Count > 1)
-            {
-                varDeclName = null;
-                localDeclaration = null;
-                return false;
-            }
-
-            varDeclName = syntaxFacts.GetIdentifierOfVariableDeclarator(declarator).ValueText;
-            return true;
         }
 
         /// <summary>
@@ -359,597 +300,6 @@ namespace Microsoft.CodeAnalysis.IntroduceVariable
             }
 
             return methodCallSites;
-        }
-
-        /// <summary>
-        /// If the parameter is optional and the invocation does not specify the parameter, then
-        /// a named argument needs to be introduced.
-        /// </summary>
-        private static SeparatedSyntaxList<SyntaxNode> AddArgumentToArgumentList(
-            SeparatedSyntaxList<SyntaxNode> invocationArguments, SyntaxGenerator generator,
-            SyntaxNode newArgumentExpression, string parameterName, int insertionIndex, bool named)
-        {
-            var argument = named
-                ? generator.Argument(parameterName, RefKind.None, newArgumentExpression)
-                : generator.Argument(newArgumentExpression);
-            return invocationArguments.Insert(insertionIndex, argument);
-        }
-
-        private static bool ShouldArgumentBeNamed(Compilation compilation, ISemanticFactsService semanticFacts,
-            SemanticModel semanticModel, SeparatedSyntaxList<SyntaxNode> invocationArguments, int methodInsertionIndex,
-            CancellationToken cancellationToken)
-        {
-            var invocationInsertIndex = 0;
-            foreach (var invocationArgument in invocationArguments)
-            {
-                var argumentParameter = semanticFacts.FindParameterForArgument(semanticModel, invocationArgument, cancellationToken);
-                if (argumentParameter is not null && ShouldParameterBeSkipped(compilation, argumentParameter))
-                {
-                    invocationInsertIndex++;
-                }
-                else
-                {
-                    break;
-                }
-            }
-
-            return invocationInsertIndex < methodInsertionIndex;
-        }
-
-        /// <summary>
-        /// Goes through the parameters of the original method to get the location that the parameter
-        /// and argument should be introduced.
-        /// </summary>
-        private static int GetInsertionIndex(Compilation compilation, IMethodSymbol methodSymbol,
-            ISyntaxFactsService syntaxFacts, SyntaxNode methodDeclaration)
-        {
-            var parameterList = syntaxFacts.GetParameterList(methodDeclaration);
-            Contract.ThrowIfNull(parameterList);
-            var insertionIndex = 0;
-
-            foreach (var parameterSymbol in methodSymbol.Parameters)
-            {
-                // Want to skip optional parameters, params parameters, and CancellationToken since they should be at
-                // the end of the list.
-                if (ShouldParameterBeSkipped(compilation, parameterSymbol))
-                {
-                    insertionIndex++;
-                }
-            }
-
-            return insertionIndex;
-        }
-
-        private static bool ShouldParameterBeSkipped(Compilation compilation, IParameterSymbol parameter)
-            => !parameter.HasExplicitDefaultValue &&
-               !parameter.IsParams &&
-               !parameter.Type.Equals(compilation.GetTypeByMetadataName(typeof(CancellationToken)?.FullName!));
-
-        /// <summary>
-        /// Gets the matches of the expression and replaces them with the identifier.
-        /// Special case for the original matching expression, if its parent is a LocalDeclarationStatement then it can
-        /// be removed because assigning the local dec variable to a parameter is repetitive. Does not need a rename
-        /// annotation since the user has already named the local declaration.
-        /// Otherwise, it needs to have a rename annotation added to it because the new parameter gets a randomly
-        /// generated name that the user can immediately change.
-        /// </summary>
-        public async Task UpdateExpressionInOriginalFunctionAsync(Document document,
-            TExpressionSyntax expression, SyntaxNode scope, string parameterName, SyntaxEditor editor,
-            bool allOccurrences, CancellationToken cancellationToken)
-        {
-            var generator = editor.Generator;
-            var matches = await FindMatchesAsync(document, expression, scope, allOccurrences, cancellationToken).ConfigureAwait(false);
-            var replacement = (TIdentifierNameSyntax)generator.IdentifierName(parameterName);
-
-            foreach (var match in matches)
-            {
-                // Special case the removal of the originating expression to either remove the local declaration
-                // or to add a rename annotation.
-                if (!match.Equals(expression))
-                {
-                    editor.ReplaceNode(match, replacement);
-                }
-                else
-                {
-                    if (ShouldRemoveVariableDeclaratorContainingExpression(document, expression, out _, out var localDeclaration))
-                    {
-                        editor.RemoveNode(localDeclaration);
-                    }
-                    else
-                    {
-                        // Found the initially selected expression. Replace it with the new name we choose, but also annotate
-                        // that name with the RenameAnnotation so a rename session is started where the user can pick their
-                        // own preferred name.
-                        replacement = (TIdentifierNameSyntax)generator.IdentifierName(generator.Identifier(parameterName)
-                            .WithAdditionalAnnotations(RenameAnnotation.Create()));
-                        editor.ReplaceNode(match, replacement);
-                    }
-                }
-            }
-        }
-
-        /// <summary>
-        /// For the trampoline case, it goes through the invocations and adds an argument which is a 
-        /// call to the extracted method.
-        /// Introduces a new method overload or new trampoline method.
-        /// Updates the original method site with a newly introduced parameter.
-        /// 
-        /// ****Trampoline Example:****
-        /// public void M(int x, int y)
-        /// {
-        ///     int f = [|x * y|];
-        ///     Console.WriteLine(f);
-        /// }
-        /// 
-        /// public void InvokeMethod()
-        /// {
-        ///     M(5, 6);
-        /// }
-        /// 
-        /// ---------------------------------------------------->
-        /// 
-        /// public int GetF(int x, int y) // Generated method
-        /// {
-        ///     return x * y;
-        /// }
-        /// 
-        /// public void M(int x, int y, int f)
-        /// {
-        ///     Console.WriteLine(f);
-        /// }
-        /// 
-        /// public void InvokeMethod()
-        /// {
-        ///     M(5, 6, GetF(5, 6)); //Fills in with call to generated method
-        /// }
-        /// 
-        /// -----------------------------------------------------------------------
-        /// ****Overload Example:****
-        /// public void M(int x, int y)
-        /// {
-        ///     int f = [|x * y|];
-        ///     Console.WriteLine(f);
-        /// }
-        /// 
-        /// public void InvokeMethod()
-        /// {
-        ///     M(5, 6);
-        /// }
-        /// 
-        /// ---------------------------------------------------->
-        /// 
-        /// public void M(int x, int y) // Generated overload
-        /// {
-        ///     M(x, y, x * y)
-        /// }
-        /// 
-        /// public void M(int x, int y, int f)
-        /// {
-        ///     Console.WriteLine(f);
-        /// }
-        /// 
-        /// public void InvokeMethod()
-        /// {
-        ///     M(5, 6);
-        /// }
-        /// </summary>
-        private async Task<SyntaxNode> ModifyDocumentInvocationsTrampolineOverloadAndIntroduceParameterAsync(Compilation compilation,
-            Document originalDocument, Document currentDocument,
-            List<SyntaxNode> invocations, Dictionary<TIdentifierNameSyntax, IParameterSymbol> mappingDictionary,
-            IMethodSymbol methodSymbol, SyntaxNode containingMethod, int insertionIndex,
-            bool allOccurrences, string parameterName, TExpressionSyntax expression, IntroduceParameterCodeActionKind selectedCodeAction,
-            CancellationToken cancellationToken)
-        {
-            var generator = SyntaxGenerator.GetGenerator(currentDocument);
-            var semanticFacts = currentDocument.GetRequiredLanguageService<ISemanticFactsService>();
-            var invocationSemanticModel = await currentDocument.GetRequiredSemanticModelAsync(cancellationToken).ConfigureAwait(false);
-
-            var root = await currentDocument.GetRequiredSyntaxRootAsync(cancellationToken).ConfigureAwait(false);
-            var editor = new SyntaxEditor(root, generator);
-            var syntaxFacts = currentDocument.GetRequiredLanguageService<ISyntaxFactsService>();
-
-            // Creating a new method name by concatenating the parameter name that has been upper-cased.
-            var newMethodIdentifier = "Get" + parameterName.ToPascalCase();
-            var validParameters = methodSymbol.Parameters.Intersect(mappingDictionary.Values).ToImmutableArray();
-
-            if (selectedCodeAction is IntroduceParameterCodeActionKind.Trampoline)
-            {
-                // Creating an empty map here to reuse so that we do not create a new dictionary for
-                // every single invocation.
-                var parameterToArgumentMap = new Dictionary<IParameterSymbol, int>();
-                foreach (var invocation in invocations)
-                {
-                    var argumentListSyntax = syntaxFacts.GetArgumentListOfInvocationExpression(invocation);
-                    editor.ReplaceNode(argumentListSyntax, (currentArgumentListSyntax, _) =>
-                    {
-                        return GenerateNewArgumentListSyntaxForTrampoline(compilation, syntaxFacts, semanticFacts, invocationSemanticModel, generator,
-                            parameterToArgumentMap, currentArgumentListSyntax,
-                            argumentListSyntax, invocation, validParameters, parameterName, newMethodIdentifier, insertionIndex, cancellationToken);
-                    });
-                }
-            }
-
-            // If you are at the original document, then also introduce the new method and introduce the parameter.
-            if (currentDocument.Id == originalDocument.Id)
-            {
-                var newMethodNode = selectedCodeAction is IntroduceParameterCodeActionKind.Trampoline
-                    ? await ExtractMethodAsync(originalDocument, expression, methodSymbol, validParameters,
-                    newMethodIdentifier, generator, cancellationToken).ConfigureAwait(false)
-                    : await GenerateNewMethodOverloadAsync(originalDocument, expression, methodSymbol, insertionIndex, generator, cancellationToken).ConfigureAwait(false);
-                editor.InsertBefore(containingMethod, newMethodNode);
-
-                await UpdateExpressionInOriginalFunctionAsync(originalDocument, expression, containingMethod,
-                    parameterName, editor, allOccurrences, cancellationToken).ConfigureAwait(false);
-                var parameterType = await GetTypeOfExpressionAsync(originalDocument, expression, cancellationToken).ConfigureAwait(false);
-                var parameter = generator.ParameterDeclaration(parameterName, generator.TypeExpression(parameterType));
-                editor.InsertParameter(containingMethod, insertionIndex, parameter);
-            }
-
-            return editor.GetChangedRoot();
-
-            // Adds an argument which is an invocation of the newly created method to the callsites
-            // of the method invocations where a parameter was added.
-            // Example:
-            // public void M(int x, int y)
-            // {
-            //     int f = [|x * y|];
-            //     Console.WriteLine(f);
-            // }
-            // 
-            // public void InvokeMethod()
-            // {
-            //     M(5, 6);
-            // }
-            //
-            // ---------------------------------------------------->
-            // 
-            // public int GetF(int x, int y)
-            // {
-            //     return x * y;
-            // }
-            // 
-            // public void M(int x, int y)
-            // {
-            //     int f = x * y;
-            //     Console.WriteLine(f);
-            // }
-            //
-            // public void InvokeMethod()
-            // {
-            //     M(5, 6, GetF(5, 6)); // This is the generated invocation which is a new argument at the call site
-            // }
-            SyntaxNode GenerateNewArgumentListSyntaxForTrampoline(Compilation compilation, ISyntaxFactsService syntaxFacts, ISemanticFactsService semanticFacts,
-                SemanticModel invocationSemanticModel, SyntaxGenerator generator, Dictionary<IParameterSymbol, int> parameterToArgumentMap,
-                SyntaxNode currentArgumentListSyntax, SyntaxNode argumentListSyntax,
-                SyntaxNode invocation, ImmutableArray<IParameterSymbol> validParameters, string parameterName, string newMethodIdentifier,
-                int insertionIndex, CancellationToken cancellationToken)
-            {
-                var invocationArguments = syntaxFacts.GetArgumentsOfArgumentList(argumentListSyntax);
-                parameterToArgumentMap.Clear();
-                MapParameterToArgumentsAtInvocation(semanticFacts, parameterToArgumentMap, invocationArguments, invocationSemanticModel, cancellationToken);
-                var currentInvocationArguments = syntaxFacts.GetArgumentsOfArgumentList(currentArgumentListSyntax);
-                var requiredArguments = new List<SyntaxNode>();
-
-                foreach (var parameterSymbol in validParameters)
-                {
-                    if (parameterToArgumentMap.TryGetValue(parameterSymbol, out var index))
-                    {
-                        requiredArguments.Add(currentInvocationArguments[index]);
-                    }
-                }
-
-                var conditionalRoot = syntaxFacts.GetRootConditionalAccessExpression(invocation);
-                var named = ShouldArgumentBeNamed(compilation, semanticFacts, invocationSemanticModel, invocationArguments, insertionIndex, cancellationToken);
-                var newMethodInvocation = GenerateNewMethodInvocation(syntaxFacts, generator, invocation, requiredArguments, newMethodIdentifier);
-
-                SeparatedSyntaxList<SyntaxNode> allArguments;
-                if (conditionalRoot is null)
-                {
-                    allArguments = AddArgumentToArgumentList(currentInvocationArguments, generator, newMethodInvocation, parameterName, insertionIndex, named);
-                }
-                else
-                {
-                    // Conditional Access expressions are parents of invocations, so it is better to just replace the
-                    // invocation in place then rebuild the tree structure.
-                    var expressionsWithConditionalAccessors = conditionalRoot.ReplaceNode(invocation, newMethodInvocation);
-                    allArguments = AddArgumentToArgumentList(currentInvocationArguments, generator, expressionsWithConditionalAccessors, parameterName, insertionIndex, named);
-                }
-
-                return UpdateArgumentListSyntax(currentArgumentListSyntax, allArguments);
-            }
-        }
-
-        private static SyntaxNode GenerateNewMethodInvocation(ISyntaxFactsService syntaxFacts, SyntaxGenerator generator, SyntaxNode invocation, List<SyntaxNode> arguments, string newMethodIdentifier)
-        {
-            var methodName = generator.IdentifierName(newMethodIdentifier);
-            var fullExpression = syntaxFacts.GetExpressionOfInvocationExpression(invocation);
-            if (syntaxFacts.IsAnyMemberAccessExpression(fullExpression))
-            {
-                var receiverExpression = syntaxFacts.GetExpressionOfMemberAccessExpression(fullExpression);
-                methodName = generator.MemberAccessExpression(receiverExpression, newMethodIdentifier);
-            }
-            else if (syntaxFacts.IsMemberBindingExpression(fullExpression))
-            {
-                methodName = generator.MemberBindingExpression(generator.IdentifierName(newMethodIdentifier));
-            }
-
-            return generator.InvocationExpression(methodName, arguments);
-        }
-
-        /// <summary>
-        /// Generates a method declaration containing a return expression of the highlighted expression.
-        /// Example:
-        /// public void M(int x, int y)
-        /// {
-        ///     int f = [|x * y|];
-        /// }
-        /// 
-        /// ---------------------------------------------------->
-        /// 
-        /// public int GetF(int x, int y)
-        /// {
-        ///     return x * y;
-        /// }
-        /// 
-        /// public void M(int x, int y)
-        /// {
-        ///     int f = x * y;
-        /// }
-        /// </summary>
-        private static async Task<SyntaxNode> ExtractMethodAsync(Document document, TExpressionSyntax expression,
-            IMethodSymbol methodSymbol, ImmutableArray<IParameterSymbol> validParameters, string newMethodIdentifier, SyntaxGenerator generator, CancellationToken cancellationToken)
-        {
-            // Remove trivia so the expression is in a single line and does not affect the spacing of the following line
-            var returnStatement = generator.ReturnStatement(expression.WithoutTrivia());
-            var typeSymbol = await GetTypeOfExpressionAsync(document, expression, cancellationToken).ConfigureAwait(false);
-            var newMethodDeclaration = await CreateMethodDeclarationAsync(document, methodSymbol, expression, returnStatement,
-                validParameters, newMethodIdentifier, typeSymbol, isTrampoline: true, cancellationToken).ConfigureAwait(false);
-            return newMethodDeclaration;
-        }
-
-        /// <summary>
-        /// Generates a method declaration containing a call to the method that introduced the parameter.
-        /// Example:
-        /// 
-        /// ***This is an intermediary step in which the original function has not be updated yet
-        /// public void M(int x, int y)
-        /// {
-        ///     int f = [|x * y|];
-        /// }
-        /// 
-        /// ---------------------------------------------------->
-        /// 
-        /// public void M(int x, int y) // Generated overload
-        /// {
-        ///     M(x, y, x * y);
-        /// }
-        /// 
-        /// public void M(int x, int y) // Original function (which will be mutated in a later step)
-        /// {
-        ///     int f = x * y;
-        /// }
-        /// </summary>
-        private static async Task<SyntaxNode> GenerateNewMethodOverloadAsync(Document document, TExpressionSyntax expression,
-            IMethodSymbol methodSymbol, int insertionIndex, SyntaxGenerator generator, CancellationToken cancellationToken)
-        {
-            // Need the parameters from the original function as arguments for the invocation
-            var arguments = generator.CreateArguments(methodSymbol.Parameters);
-
-            // Remove trivia so the expression is in a single line and does not affect the spacing of the following line
-            arguments = arguments.Insert(insertionIndex, generator.Argument(expression.WithoutTrivia()));
-            var memberName = methodSymbol.IsGenericMethod
-                ? generator.GenericName(methodSymbol.Name, methodSymbol.TypeArguments)
-                : generator.IdentifierName(methodSymbol.Name);
-            var invocation = generator.InvocationExpression(memberName, arguments);
-
-            var newStatement = methodSymbol.ReturnsVoid
-               ? generator.ExpressionStatement(invocation)
-               : generator.ReturnStatement(invocation);
-
-            var newMethodDeclaration = await CreateMethodDeclarationAsync(document, methodSymbol, expression, newStatement,
-                validParameters: null, newMethodIdentifier: null, typeSymbol: null, isTrampoline: false, cancellationToken).ConfigureAwait(false);
-            return newMethodDeclaration;
-        }
-
-        private static async Task<SyntaxNode> CreateMethodDeclarationAsync(Document document, IMethodSymbol methodSymbol,
-            TExpressionSyntax expression, SyntaxNode newStatement, ImmutableArray<IParameterSymbol>? validParameters,
-            string? newMethodIdentifier, ITypeSymbol? typeSymbol, bool isTrampoline, CancellationToken cancellationToken)
-        {
-            var codeGenerationService = document.GetRequiredLanguageService<ICodeGenerationService>();
-            var options = await document.GetOptionsAsync(cancellationToken).ConfigureAwait(false);
-
-            var newMethod = isTrampoline
-                ? CodeGenerationSymbolFactory.CreateMethodSymbol(methodSymbol, name: newMethodIdentifier, parameters: validParameters, statements: ImmutableArray.Create(newStatement), returnType: typeSymbol)
-                : CodeGenerationSymbolFactory.CreateMethodSymbol(methodSymbol, statements: ImmutableArray.Create(newStatement), containingType: methodSymbol.ContainingType);
-            var newMethodDeclaration = codeGenerationService.CreateMethodDeclaration(newMethod, options: new CodeGenerationOptions(options: options, parseOptions: expression.SyntaxTree.Options));
-            return newMethodDeclaration;
-        }
-
-        /// <summary>
-        /// This method goes through all the invocation sites and adds a new argument with the expression to be added.
-        /// It also introduces a parameter at the original method site.
-        /// 
-        /// Example:
-        /// public void M(int x, int y)
-        /// {
-        ///     int f = [|x * y|];
-        /// }
-        /// 
-        /// public void InvokeMethod()
-        /// {
-        ///     M(5, 6);
-        /// }
-        /// 
-        /// ---------------------------------------------------->
-        /// 
-        /// public void M(int x, int y, int f) // parameter gets introduced
-        /// {
-        /// }
-        /// 
-        /// public void InvokeMethod()
-        /// {
-        ///     M(5, 6, 5 * 6); // argument gets added to callsite
-        /// }
-        /// </summary>
-        private async Task<SyntaxNode> ModifyDocumentInvocationsAndIntroduceParameterAsync(Compilation compilation,
-            Document originalDocument, Document document, Dictionary<TIdentifierNameSyntax, IParameterSymbol> mappingDictionary,
-            SyntaxNode containingMethod, TExpressionSyntax expression,
-            bool allOccurrences, string parameterName, int insertionIndex,
-            List<SyntaxNode> invocations, CancellationToken cancellationToken)
-        {
-            var semanticFacts = document.GetRequiredLanguageService<ISemanticFactsService>();
-            var syntaxFacts = document.GetRequiredLanguageService<ISyntaxFactsService>();
-            var root = await document.GetRequiredSyntaxRootAsync(cancellationToken).ConfigureAwait(false);
-            var generator = SyntaxGenerator.GetGenerator(document);
-            var editor = new SyntaxEditor(root, generator);
-            var invocationSemanticModel = await document.GetRequiredSemanticModelAsync(cancellationToken).ConfigureAwait(false);
-            var parameterToArgumentMap = new Dictionary<IParameterSymbol, int>();
-            foreach (var invocation in invocations)
-            {
-                var expressionEditor = new SyntaxEditor(expression, generator);
-
-                var argumentListSyntax = invocation is TObjectCreationExpressionSyntax
-                    ? syntaxFacts.GetArgumentListOfObjectCreationExpression(invocation)
-                    : syntaxFacts.GetArgumentListOfInvocationExpression(invocation);
-
-                var invocationArguments = syntaxFacts.GetArgumentsOfArgumentList(argumentListSyntax);
-                parameterToArgumentMap.Clear();
-                MapParameterToArgumentsAtInvocation(semanticFacts, parameterToArgumentMap, invocationArguments, invocationSemanticModel, cancellationToken);
-
-                if (argumentListSyntax is not null)
-                {
-                    editor.ReplaceNode(argumentListSyntax, (currentArgumentListSyntax, _) =>
-                    {
-                        var updatedInvocationArguments = syntaxFacts.GetArgumentsOfArgumentList(currentArgumentListSyntax);
-                        var updatedExpression = CreateNewArgumentExpression(expressionEditor,
-                            syntaxFacts, mappingDictionary, parameterToArgumentMap, updatedInvocationArguments);
-                        var named = ShouldArgumentBeNamed(compilation, semanticFacts, invocationSemanticModel, invocationArguments, insertionIndex, cancellationToken);
-                        var allArguments = AddArgumentToArgumentList(updatedInvocationArguments, generator,
-                            updatedExpression.WithAdditionalAnnotations(Formatter.Annotation), parameterName, insertionIndex, named);
-                        return UpdateArgumentListSyntax(currentArgumentListSyntax, allArguments);
-                    });
-                }
-            }
-
-            // If you are at the original document, then also introduce the new method and introduce the parameter.
-            if (document.Id == originalDocument.Id)
-            {
-                await UpdateExpressionInOriginalFunctionAsync(originalDocument, expression, containingMethod,
-                    parameterName, editor, allOccurrences, cancellationToken).ConfigureAwait(false);
-                var parameterType = await GetTypeOfExpressionAsync(document, expression, cancellationToken).ConfigureAwait(false);
-                var parameter = generator.ParameterDeclaration(name: parameterName, type:
-                    generator.TypeExpression(parameterType));
-                editor.InsertParameter(containingMethod, insertionIndex, parameter);
-            }
-
-            return editor.GetChangedRoot();
-        }
-
-        private static async Task<ITypeSymbol> GetTypeOfExpressionAsync(Document document, TExpressionSyntax expression, CancellationToken cancellationToken)
-        {
-            var semanticModel = await document.GetRequiredSemanticModelAsync(cancellationToken).ConfigureAwait(false);
-            var typeSymbol = semanticModel.GetTypeInfo(expression, cancellationToken).ConvertedType ?? semanticModel.Compilation.ObjectType;
-            return typeSymbol;
-        }
-        /// <summary>
-        /// This method iterates through the variables in the expression and maps the variables back to the parameter
-        /// it is associated with. It then maps the parameter back to the argument at the invocation site and gets the
-        /// index to retrieve the updated arguments at the invocation.
-        /// </summary>
-        private TExpressionSyntax CreateNewArgumentExpression(SyntaxEditor editor, ISyntaxFactsService syntaxFacts,
-            Dictionary<TIdentifierNameSyntax, IParameterSymbol> mappingDictionary,
-            Dictionary<IParameterSymbol, int> parameterToArgumentMap,
-            SeparatedSyntaxList<SyntaxNode> updatedInvocationArguments)
-        {
-            foreach (var (variable, mappedParameter) in mappingDictionary)
-            {
-                var parameterMapped = parameterToArgumentMap.TryGetValue(mappedParameter, out var index);
-                if (parameterMapped)
-                {
-                    var updatedInvocationArgument = updatedInvocationArguments[index];
-                    var argumentExpression = syntaxFacts.GetExpressionOfArgument(updatedInvocationArgument);
-                    var parenthesizedArgumentExpression = editor.Generator.AddParentheses(argumentExpression, includeElasticTrivia: false);
-                    editor.ReplaceNode(variable, parenthesizedArgumentExpression);
-                }
-                else if (mappedParameter.HasExplicitDefaultValue)
-                {
-                    var generatedExpression = GenerateExpressionFromOptionalParameter(mappedParameter);
-                    var parenthesizedGeneratedExpression = editor.Generator.AddParentheses(generatedExpression, includeElasticTrivia: false);
-                    editor.ReplaceNode(variable, parenthesizedGeneratedExpression);
-                }
-            }
-
-            return (TExpressionSyntax)editor.GetChangedRoot();
-        }
-
-        private static void MapParameterToArgumentsAtInvocation(
-            ISemanticFactsService semanticFacts, Dictionary<IParameterSymbol, int> mapping, SeparatedSyntaxList<SyntaxNode> arguments,
-            SemanticModel invocationSemanticModel, CancellationToken cancellationToken)
-        {
-            for (var i = 0; i < arguments.Count; i++)
-            {
-                var argumentParameter = semanticFacts.FindParameterForArgument(invocationSemanticModel, arguments[i], cancellationToken);
-                if (argumentParameter is not null)
-                {
-                    mapping[argumentParameter] = i;
-                }
-            }
-        }
-
-        /// <summary>
-        /// Ties the identifiers within the expression back to their associated parameter.
-        /// </summary>
-        public static async Task<Dictionary<TIdentifierNameSyntax, IParameterSymbol>> MapExpressionToParametersAsync(
-            Document document, TExpressionSyntax expression, CancellationToken cancellationToken)
-        {
-            var nameToParameterDict = new Dictionary<TIdentifierNameSyntax, IParameterSymbol>();
-            var variablesInExpression = expression.DescendantNodes().OfType<TIdentifierNameSyntax>();
-            var semanticModel = await document.GetRequiredSemanticModelAsync(cancellationToken).ConfigureAwait(false);
-
-            foreach (var variable in variablesInExpression)
-            {
-                var symbol = semanticModel.GetSymbolInfo(variable, cancellationToken).Symbol;
-                if (symbol is IParameterSymbol parameterSymbol)
-                {
-                    nameToParameterDict.Add(variable, parameterSymbol);
-                }
-            }
-
-            return nameToParameterDict;
-        }
-
-        /// <summary>
-        /// Finds the matches of the expression within the same block.
-        /// </summary>
-        protected static async Task<IEnumerable<TExpressionSyntax>> FindMatchesAsync(Document document,
-            TExpressionSyntax expression, SyntaxNode withinNode,
-            bool allOccurrences, CancellationToken cancellationToken)
-        {
-            if (!allOccurrences)
-            {
-                return SpecializedCollections.SingletonEnumerable(expression);
-            }
-
-            var syntaxFacts = document.GetRequiredLanguageService<ISyntaxFactsService>();
-            var originalSemanticModel = await document.GetRequiredSemanticModelAsync(cancellationToken).ConfigureAwait(false);
-
-            var matches = from nodeInCurrent in withinNode.DescendantNodesAndSelf().OfType<TExpressionSyntax>()
-                          where NodeMatchesExpression(originalSemanticModel, expression, nodeInCurrent, cancellationToken)
-                          select nodeInCurrent;
-            return matches;
-        }
-
-        private static bool NodeMatchesExpression(SemanticModel originalSemanticModel,
-            TExpressionSyntax expression, TExpressionSyntax currentNode, CancellationToken cancellationToken)
-        {
-            cancellationToken.ThrowIfCancellationRequested();
-            if (currentNode == expression)
-            {
-                return true;
-            }
-
-            return SemanticEquivalence.AreEquivalent(
-                originalSemanticModel, originalSemanticModel, expression, currentNode);
         }
 
         private class MyCodeAction : SolutionChangeAction
